@@ -5,7 +5,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.6 <4.0.0'
+  sdk: '>=3.4.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -15,11 +15,10 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
-  intl: ^0.18.0
-  google_fonts: ^5.1.0
+  intl: ^0.19.0
   flutter_svg: ^2.0.7
   sembast: ^3.5.0+1
-  uuid: ^3.0.7
+  uuid: ^4.4.0
   path: ^1.8.3
   path_provider: ^2.0.15
   firebase_core: ^2.15.1
@@ -34,11 +33,23 @@ dependencies:
   image_picker: ^1.0.4
   flutter_staggered_grid_view: ^0.7.0
   http_parser: ^4.0.2
-
+  duration_time_picker:
+    git: 
+      url: https://github.com/cledsonv/duration_time_picker.git
+  google_fonts: ^6.2.1
+  flutter_quill:
+    git: 
+      url: https://github.com/cledsonv/flutter-quill.git
+  vsc_quill_delta_to_html: ^1.0.4
+  flutter_widget_from_html_core: ^0.14.11
+  dotted_border: ^2.1.0
+  infinite_scroll_pagination: ^4.0.0
+dependency_overrides:
+  intl: ^0.19.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
 flutter:
   generate: true
