@@ -1,17 +1,14 @@
-import 'package:app_receitas/src/core/widgets/cookie_text.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text.dart';
 import 'package:app_receitas/src/feactures/recipes/presenter/ui/atomic/custom_container.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 
 class ContainerCreateInfo extends StatelessWidget {
-  final String title;
+  final String? title;
   final String iconSvg;
   final Widget child;
   const ContainerCreateInfo(
-      {super.key,
-      required this.title,
-      required this.iconSvg,
-      required this.child});
+      {super.key, this.title, required this.iconSvg, required this.child});
 
   @override
   Widget build(BuildContext context) {
@@ -23,7 +20,7 @@ class ContainerCreateInfo extends StatelessWidget {
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: [
               CookieText(
-                text: title,
+                text: title ?? '',
                 typography: CookieTypography.button,
               ),
               SvgPicture.asset(iconSvg),
