@@ -1,56 +1,85 @@
 import 'dart:convert';
 import 'dart:developer';
 
+import 'package:app_receitas/src/core/library/api_client.dart';
+import 'package:app_receitas/src/core/services/api/response_request_entity.dart';
+import 'package:app_receitas/src/core/services/api/response_request_mapper.dart';
 import 'package:dio/dio.dart';
 
-class DioClient {
+class DioClient extends ApiClient {
   final Dio _dio = Dio();
-  Future<Response<T>> put<T>(
+  final ResponseRequestMapper _requestMapper;
+
+  DioClient(this._requestMapper);
+
+  @override
+  Future<ResponseRequestEntity> put<T>(
     String url, {
     required Map<String, dynamic> body,
     required Map<String, dynamic> headers,
-  }) {
-    _log(method: 'put', path: url, headers: headers, body: body);
-    return _dio.put<T>(
+    bool isformData = false,
+  }) async {
+    dynamic data = body;
+    if (isformData) {
+      _log(method: 'put', path: url, headers: headers);
+      data = FormData.fromMap(data);
+    } else {
+      _log(method: 'put', path: url, headers: headers, body: body);
+    }
+    final response = await _dio.put(
       url,
-      data: body,
+      data: data,
       options: Options(headers: headers),
     );
+    return _requestMapper.fromDio(response);
   }
 
-  Future<Response<T>> delete<T>(
+  @override
+  Future<ResponseRequestEntity> delete<T>(
     String url, {
     required Map<String, dynamic> headers,
-  }) {
+  }) async {
     _log(method: 'delete', path: url, headers: headers);
-    return _dio.delete<T>(
+    final response = await _dio.delete<T>(
       url,
       options: Options(headers: headers),
     );
+    return _requestMapper.fromDio(response);
   }
 
-  Future<Response<T>> get<T>(
+  @override
+  Future<ResponseRequestEntity> get<T>(
     String url, {
     required Map<String, dynamic> headers,
-  }) {
+  }) async {
     _log(method: 'get', path: url, headers: headers);
-    return _dio.get<T>(
+    final response = await _dio.get<T>(
       url,
       options: Options(headers: headers),
     );
+    return _requestMapper.fromDio(response);
   }
 
-  Future<Response<T>> post<T>(
+  @override
+  Future<ResponseRequestEntity> post<T>(
     String url, {
     required Map<String, dynamic> body,
     required Map<String, dynamic> headers,
-  }) {
-    _log(method: 'post', path: url, headers: headers, body: body);
-    return _dio.post<T>(
+    bool isformData = false,
+  }) async {
+    dynamic data = body;
+    if (isformData) {
+      _log(method: 'post', path: url, headers: headers);
+      data = FormData.fromMap(data);
+    } else {
+      _log(method: 'post', path: url, headers: headers, body: body);
+    }
+    final response = await _dio.post<T>(
       url,
       data: body,
       options: Options(headers: headers),
     );
+    return _requestMapper.fromDio(response);
   }
 
   void _log({
