@@ -1,4 +1,4 @@
-import 'package:app_receitas/src/core/services/api_response/api_error/api_error.dart';
+import 'package:app_receitas/src/core/services/api/api_error/api_error.dart';
 
 class ApiInternalServerError extends ApiError {
   @override
