@@ -1,5 +1,5 @@
-import 'package:app_receitas/src/core/widgets/cookie_text.dart';
-import 'package:app_receitas/src/core/widgets/cookie_text_button.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text_button.dart';
 import 'package:flutter/material.dart';
 
 class CookieDialog extends StatelessWidget {
