@@ -1,8 +1,8 @@
 import 'package:app_receitas/src/core/global/global_variables.dart';
-import 'package:app_receitas/src/core/widgets/cookie_button.dart';
-import 'package:app_receitas/src/core/widgets/cookie_text.dart';
-import 'package:app_receitas/src/core/widgets/cookie_text_button.dart';
-import 'package:app_receitas/src/core/widgets/cookie_text_field.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_button.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text_button.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text_field.dart';
 import 'package:app_receitas/src/feactures/auth/presenter/controllers/auth_controller.dart';
 import 'package:app_receitas/src/feactures/auth/presenter/ui/organisms/custom_screen.dart';
 import 'package:app_receitas/src/feactures/auth/presenter/ui/pages/forget_password_page.dart';
@@ -112,7 +112,7 @@ class _LoginPageState extends State<LoginPage> {
           label: AppLocalizations.of(context)!.loginButton,
           onPressed: () async {
             var result = await ct.loginFirebase();
-            if (!result && mounted) {
+            if (!result && context.mounted) {
               final snackBar = SnackBar(
                 content: CookieText(
                   text: AppLocalizations.of(context)!.loginTextSnack,
