@@ -3,5 +3,5 @@ import 'package:app_receitas/src/feactures/profile/domain/entities/profile_entit
 abstract class ProfileRepository {
   Future<ProfileEntity> getProfile(String userID);
   Future<void> updateProfile(ProfileEntity profileEntity);
-  Future<void> updateImageProfile(String userID, String? imagePath);
+  Future<void> updateImageProfile(String userID, String imagePath);
 }
