@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:app_receitas/src/core/widgets/cookie_page.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_page.dart';
 import 'package:app_receitas/src/feactures/profile/domain/entities/profile_entity.dart';
 import 'package:app_receitas/src/feactures/profile/domain/repositories/profile_repository.dart';
 import 'package:flutter/material.dart';
@@ -28,7 +28,7 @@ class EditProfileController extends ChangeNotifier {
 
   Future<void> updateImageProfile() async {
     if (image == null) return;
-    await _repository.updateImageProfile(profile!.userID, image?.path);
+    await _repository.updateImageProfile(profile!.userID, image!.path);
   }
 
   Future<void> pickImageLogo() async {
