@@ -1,19 +1,16 @@
 import 'package:app_receitas/src/core/global/global_variables.dart';
-import 'package:app_receitas/src/core/library/dio_client.dart';
-import 'package:app_receitas/src/core/services/api_response/api_response.dart';
+import 'package:app_receitas/src/core/services/api/api_recipes.dart';
 import 'package:app_receitas/src/feactures/recipes/data/database/recipe_database.dart';
 import 'package:app_receitas/src/feactures/recipes/data/mappers/recipe_mapper.dart';
 import 'package:app_receitas/src/feactures/recipes/domain/entities/recipe_entity.dart';
 
 class RecipeDatabaseImp implements RecipeDatabase {
-  final DioClient dio;
   final RecipeMapper _mapper;
-  final ApiResponse _apiResponse;
+  final ApiRecipes _apiRecipes;
 
   RecipeDatabaseImp(
-    this.dio,
     this._mapper,
-    this._apiResponse,
+    this._apiRecipes,
   );
 
   String url = Global.dnsApi;
@@ -21,39 +18,20 @@ class RecipeDatabaseImp implements RecipeDatabase {
 
   @override
   Future<void> createRecipe(RecipeEntity recipe) async {
-    final result = await dio.post(
-      '$url/$path',
-      body: _mapper.toMap(recipe),
-      headers: {
-        'Authorization': Global.token,
-      },
-    );
-    _apiResponse.handleResponse(result);
+    await _apiRecipes.post(path: '$url/$path', body: _mapper.toMap(recipe));
   }
 
   @override
   Future<RecipeEntity> getRecipe(String id) async {
-    final response = await dio.get(
-      '$url/$path/$id',
-      headers: {
-        'Authorization': Global.token,
-      },
-    );
-    final result = _apiResponse.handleResponse(response);
+    final result = await _apiRecipes.get(path: '$url/$path/$id');
 
     return _mapper.fromMap(result);
   }
 
   @override
   Future<List<RecipeEntity>> listRecipe() async {
-    final response = await dio.get(
-      '$url/$path',
-      headers: {
-        'Authorization': Global.token,
-      },
-    );
-    final result = _apiResponse.handleResponse(response);
-    
+    final result = await _apiRecipes.get(path: '$url/$path');
+
     return result.map<RecipeEntity>((e) => _mapper.fromMap(e)).toList();
   }
 }
