@@ -1,4 +1,4 @@
-import 'package:app_receitas/src/core/widgets/cookie_text.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text.dart';
 import 'package:flutter/material.dart';
 
 class RecipeSearchContainer extends StatelessWidget {
