@@ -1,6 +1,6 @@
-import 'package:app_receitas/src/core/widgets/cookie_text.dart';
-import 'package:app_receitas/src/core/widgets/cookie_text_button.dart';
-import 'package:app_receitas/src/core/widgets/cookie_text_field_search.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text_button.dart';
+import 'package:app_receitas/src/core/widgets/feactures/cookie_text_field_search.dart';
 import 'package:app_receitas/src/feactures/search/presenter/ui/pages/filter_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
